create general logo component
