+++ packages/core/src/Footer/index.js
@@ -33,7 +33,6 @@ export default Footer;
 // Expose components
 export { default as FooterAbout } from "./About";
 export { default as FooterCopyright } from "./Copyright";
-export { default as FooterLogo } from "./Logo";
 export { default as FooterInitiativeLogo } from "./Initiative";
 export { default as FooterLegalLinks } from "./LegalLinks";
 export { default as FooterQuickLinks } from "./QuickLinks";
